refactor(types): separator
